chore(tests): extract cross-test shared helpers into tests/_support/ (incremental Step 1)
Step 1 of incremental tests/ organization. Extracts the helpers that test
modules imported from sibling test files into a stable, location-independent
support package, and repoints the importers. Bulk migration of the ~700 flat
test files is DEFERRED to a later staged step; issue-number suffixes are KEPT
(owner).

Problem: 28 test files imported shared helpers via `from tests.test_<mod>
import ...`. That binds a helper to a *test file's* location — it breaks the
moment the file moves (the deferred migration) and historically only resolved
under full-suite namespace-package collection (bare `pytest <file>` from a
non-root cwd failed with `ModuleNotFoundError: No module named 'tests'`).

Change:
- New `tests/_support/` package (real `__init__.py`) holding the shared
  helpers, moved byte-identical (not rewritten):
  - permissions.py: make_resolver
  - router_loop.py: FakeRouterHost, FakeEventLog, ScriptedLLM, text_result,
    tool_result, make_loop, EMPTY_USAGE
  - router_host_adapter.py: make_adapter + inert null_* action ports
  - session.py: make_session, push, now, synthetic_t_max
- Source modules now import their own helpers from _support (aliased to the
  original module-local names), so their test bodies are unchanged.
- 28 cross-test imports across 26 files repointed to tests._support.* .
- tests/conftest.py inserts the repo root on sys.path so tests._support
  resolves under any invocation style (bare pytest / non-root cwd / IDE
  runner), not only `python -m pytest` from the repo root.
- tests/README.md documents the layout convention (new tests → subsystem
  subdirs mirroring src/reyn/; suffixes kept; bulk migration deferred) and the
  _support shared-helper rule.

Deferred (documented in tests/README.md): two force-close scenario modules
still import scenario-specific builders from a sibling test module; their
builders are tightly coupled to single-use scaffolding, so extracting them is a
follow-up to avoid churn in the active force-close area. Both still collect in
isolation via the conftest bootstrap.

Verification: 191 impacted tests pass; isolated `--collect-only` now succeeds
under both `python -m pytest` and bare `pytest` from another cwd; full-suite
collection adds zero new errors vs origin/main (5 pre-existing fastapi
optional-dep errors unchanged); test_tier_audit.py clean (188 tests, 0
warnings/errors); ruff clean.
