+++ tests/test_1199_s31b2c2_http_get.py
@@ -16,7 +16,7 @@ import pytest
 
 from reyn.security.permissions.effective import AgentLayer, CapabilityAxis
 from reyn.security.permissions.permissions import PermissionDecl
-from tests.test_permissions import _make_resolver
+from tests._support.permissions import make_resolver as _make_resolver
 
 AX = CapabilityAxis
