+++ tests/test_a2a_router_cap_wrapup_1538.py
@@ -39,7 +39,7 @@ from reyn.runtime.errors import RouterCapExceeded
 from reyn.runtime.services.a2a_handler import A2AHandler
 from reyn.runtime.services.chain_manager import ChainManager
 from reyn.runtime.session import Session
-from tests.test_router_loop import FakeEventLog
+from tests._support.router_loop import FakeEventLog
 
 _EMPTY_USAGE = TokenUsage(prompt_tokens=10, completion_tokens=5)
