+++ tests/test_cwd_sandbox_aware_1477.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 import os
 from pathlib import Path
 
-from tests.test_router_host_adapter_invariants import _make_adapter
+from tests._support.router_host_adapter import make_adapter as _make_adapter
 
 # ── Real fake backends ───────────────────────────────────────────────────────
