+++ tests/test_force_close_trigger_1092.py
@@ -26,7 +26,7 @@ import pytest
 from reyn.llm.llm import LLMToolCallResult
 from reyn.llm.pricing import TokenUsage
 from reyn.runtime.router_loop import RouterLoop
-from tests.test_router_loop import FakeRouterHost
+from tests._support.router_loop import FakeRouterHost
 
 
 class _ForceCloseHost(FakeRouterHost):
