+++ tests/test_format_feedback_golden_1608.py
@@ -21,7 +21,7 @@ import pytest
 from reyn.llm.llm import LLMToolCallResult
 from reyn.llm.pricing import TokenUsage
 from reyn.runtime.router_loop import RouterLoop
-from tests.test_router_loop import FakeRouterHost, text_result, tool_result
+from tests._support.router_loop import FakeRouterHost, text_result, tool_result
 
 _USAGE = TokenUsage(prompt_tokens=10, completion_tokens=5)
