+++ tests/test_router_host_adapter_invariants.py
@@ -38,136 +38,54 @@ def _null_async(*args, **kwargs):
     return _inner()
 
 
-async def _null_file_read(path: str) -> dict:
-    return {"content": ""}
-
-
-async def _null_file_write(path: str, content: str) -> dict:
-    return {"path": path, "written": True}
-
-
-async def _null_file_delete(path: str) -> dict:
-    return {"path": path, "deleted": True}
-
-
-async def _null_file_list(path: str) -> dict:
-    return {"path": path, "entries": []}
-
-
-async def _null_file_regen(*, path, output_path, entry_template, header) -> dict:
-    return {"path": path, "output_path": output_path, "entries": 0}
-
-
-async def _null_mcp_list_servers() -> list:
-    return []
-
-
-async def _null_mcp_list_tools(server: str) -> list:
-    return []
-
-
-async def _null_mcp_call_tool(server: str, tool: str, args: dict) -> dict:
-    return {}
-
-
-async def _null_run_skill(spec, *, chain_id) -> dict:
-    return {"status": "finished", "data": {}}
-
-
-async def _null_spawn_skill(spec, *, chain_id) -> dict:
-    """FP-0012: spawn-ack stub for adapter tests."""
-    return {
-        "status": "spawned",
-        "run_id": "test-run-id",
-        "chain_id": chain_id,
-        "skill": spec.get("skill", ""),
-        "note": "test stub",
-    }
-
-
-async def _null_send_to_agent(*, to, request, depth, chain_id) -> None:
-    pass
-
-
-async def _null_put_outbox(msg) -> None:
-    pass
-
-
-def _null_append_history(msg) -> None:
-    pass
-
-
-async def _null_spawn_plan_task(*, plan_id, runtime, chain_id, parent_chain_id=None) -> None:
-    pass
-
-
-def _make_adapter(
-    agent_name: str = "test-agent",
-    agent_workspace_dir: Path | None = None,
-    events: EventLog | None = None,
-    memory: MemoryService | None = None,
-    delegation_list: "list[dict] | None" = None,
-    agent_replies_list: "list[str] | None" = None,
-    resolver: ModelResolver | None = None,
-    turn_budget_engine: object = None,
-    environment_backend: object = None,  # #1477: optional for sandbox-cwd tests
-) -> RouterHostAdapter:
-    """Construct a minimal RouterHostAdapter with real collaborators."""
-    if events is None:
-        events = EventLog(subscribers=[])
-    workspace = agent_workspace_dir or Path(".reyn") / "agents" / agent_name
-    if memory is None:
-        memory = MemoryService(
-            agent_workspace_dir=workspace,
-            events=events,
-            file_write=_null_file_write,
-            file_read=_null_file_read,
-            file_delete=_null_file_delete,
-            file_regenerate_index=_null_file_regen,
-        )
-    if resolver is None:
-        resolver = ModelResolver({})
-
-    _delegations = delegation_list
-    _replies = agent_replies_list
-
-    return RouterHostAdapter(
-        agent_name=agent_name,
-        agent_role="test role",
-        output_language="en",
-        allowed_skills=None,
-        allowed_mcp=None,
-        permission_resolver=None,
-        mcp_servers=None,
-        project_context="",
-        events=events,
-        resolver=resolver,
-        memory=memory,
-        journal=None,
-        agent_registry=None,
-        skill_enumerate_fn=lambda exclude: [],
-        agent_workspace_dir=workspace,
-        plan_registry_getter=lambda: None,
-        file_read=_null_file_read,
-        file_write=_null_file_write,
-        file_delete=_null_file_delete,
-        file_list_directory=_null_file_list,
-        file_regenerate_index=_null_file_regen,
-        mcp_list_servers=_null_mcp_list_servers,
-        mcp_list_tools=_null_mcp_list_tools,
-        mcp_call_tool=_null_mcp_call_tool,
-        run_skill_awaitable=_null_run_skill,
-        spawn_skill=_null_spawn_skill,
-        send_to_agent=_null_send_to_agent,
-        put_outbox=_null_put_outbox,
-        append_history=_null_append_history,
-        spawn_plan_task=_null_spawn_plan_task,
-        delegation_tracker=lambda: _delegations,
-        agent_replies_tracker=lambda: _replies,
-        turn_budget_engine=turn_budget_engine,
-        environment_backend=environment_backend,
-    )
-
+# The make_adapter builder and its inert null_* action ports now live in
+# tests/_support (stable, location-independent import path). Aliased back to the
+# original module-local names so the tests below are unchanged.
+from tests._support.router_host_adapter import (  # noqa: E402
+    make_adapter as _make_adapter,
+)
+from tests._support.router_host_adapter import (
+    null_append_history as _null_append_history,
+)
+from tests._support.router_host_adapter import (
+    null_file_delete as _null_file_delete,
+)
+from tests._support.router_host_adapter import (
+    null_file_list as _null_file_list,
+)
+from tests._support.router_host_adapter import (
+    null_file_read as _null_file_read,
+)
+from tests._support.router_host_adapter import (
+    null_file_regen as _null_file_regen,
+)
+from tests._support.router_host_adapter import (
+    null_file_write as _null_file_write,
+)
+from tests._support.router_host_adapter import (
+    null_mcp_call_tool as _null_mcp_call_tool,
+)
+from tests._support.router_host_adapter import (
+    null_mcp_list_servers as _null_mcp_list_servers,
+)
+from tests._support.router_host_adapter import (
+    null_mcp_list_tools as _null_mcp_list_tools,
+)
+from tests._support.router_host_adapter import (
+    null_put_outbox as _null_put_outbox,
+)
+from tests._support.router_host_adapter import (
+    null_run_skill as _null_run_skill,
+)
+from tests._support.router_host_adapter import (
+    null_send_to_agent as _null_send_to_agent,
+)
+from tests._support.router_host_adapter import (
+    null_spawn_plan_task as _null_spawn_plan_task,
+)
+from tests._support.router_host_adapter import (
+    null_spawn_skill as _null_spawn_skill,
+)
 
 # ---------------------------------------------------------------------------
 # Test 1: Protocol conformance (runtime_checkable isinstance)
