+++ tests/test_safety_max_iterations_fp0005.py
@@ -33,7 +33,8 @@ from reyn.llm.llm import LLMToolCallResult
 from reyn.llm.pricing import TokenUsage
 from reyn.runtime.router_loop import RouterLoop
 from reyn.user_intervention import InterventionAnswer, UserIntervention
-from tests.test_router_loop import FakeRouterHost, _ScriptedLLM, text_result
+from tests._support.router_loop import FakeRouterHost, text_result
+from tests._support.router_loop import ScriptedLLM as _ScriptedLLM
 
 _EMPTY_USAGE = TokenUsage(prompt_tokens=10, completion_tokens=5)
