+++ tests/test_turn_cancel_1468.py
@@ -25,10 +25,12 @@ import pytest
 from reyn.llm.llm import LLMToolCallResult
 from reyn.llm.pricing import TokenUsage
 from reyn.runtime.router_loop import RouterLoop
-from tests.test_router_loop import (
+from tests._support.router_loop import (
     FakeRouterHost,
-    _ScriptedLLM,
     text_result,
+)
+from tests._support.router_loop import (
+    ScriptedLLM as _ScriptedLLM,
 )
